requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
openpyxl
